+++ __init__.py
@@ -4,5 +4,5 @@ WNote - Terminal Note Taking Application
 
 from .wnote import cli
 
-__version__ = "0.5.1"
+__version__ = "0.5.2"
 __all__ = ["cli"]